+++ R/render_docs.R
@@ -1,5 +1,0 @@
-rmarkdown::render(input = "docs/documentation.Rmd",
-                  output_dir = "www"
-                  # output_format = "HTML",
-                  # output_options = "mathjax = https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-AMS_CHTML.js"
-)