+++ data-raw/create_constants.R
@@ -3,12 +3,13 @@ taxonomicBackboneMethod_options <- c("Bundled" = "bundled",
                                      "Upload" = "upload",
                                      "Kew WCVP" = "wcvp",
                                      "Syntopic Tables" = "syntopicTables")
+usethis::use_data(taxonomicBackboneMethod_options, overwrite = TRUE, internal = TRUE)
 
 # Input method options ----------------------------------------------------
 inputMethod_options <- c("Manual" = "manual",
                          "Example" = "example",
                          "Upload" = "upload")
-
+usethis::use_data(inputMethod_options, overwrite = TRUE, internal = TRUE)
 
 # Example data options ----------------------------------------------------
 example_data_options <- c("None" = "none", 
@@ -16,13 +17,14 @@ example_data_options <- c("None" = "none",
                           "Whitwell Common (M)" = "Whitwell Common",
                           "Leith Hill Place Wood (W)" = "Leith Hill Place Wood",
                           "Newborough Warren (SD)" = "Newborough Warren")
-
-
+usethis::use_data(example_data_options, overwrite = TRUE, internal = TRUE)
 
 # Data Entry Format Options -----------------------------------------------
 dataEntryFormat_options <- c("Long" = "long",
                              "Wide" = "wide",
-                             "Matrix" = "matrix")
+                             "Matrix" = "matrix",
+                             "MAVIS" = "mavis")
+usethis::use_data(dataEntryFormat_options, overwrite = TRUE, internal = TRUE)
 
 # Domin Cover -------------------------------------------------------------
 dominCoverMidPointPerc <- c("10" = 0.955,
@@ -34,8 +36,8 @@ dominCoverMidPointPerc <- c("10" = 0.955,
                             "4" = 0.07,
                             "3" = 0.04,
                             "2" = 0.025,
-                            "1" = 0.01
-)
+                            "1" = 0.01)
+usethis::use_data(dominCoverMidPointPerc, overwrite = TRUE, internal = TRUE)
 
 dominCoverVals <- c("91-100%" = 10,
                     "76-90%" = 9,
@@ -47,6 +49,7 @@ dominCoverVals <- c("91-100%" = 10,
                     "<4% (many individuals)" = 3,
                     "<4% (several individuals)" = 2,
                     "<4% (few individuals)" = 1)
+usethis::use_data(dominCoverVals, overwrite = TRUE, internal = TRUE)
 
 dominCoverValsRev <- c("10" = "91-100%",
                        "9" = "76-90%",
@@ -58,7 +61,7 @@ dominCoverValsRev <- c("10" = "91-100%",
                        "3" = "<4% (many individuals)",
                        "2" = "<4% (several individuals)",
                        "1" = "<4% (few individuals)")
-
+usethis::use_data(dominCoverValsRev, overwrite = TRUE, internal = TRUE)
 
 # Frequency Classes -------------------------------------------------------
 freqClasses_numToPerc <- list(
@@ -68,6 +71,7 @@ freqClasses_numToPerc <- list(
   "IV" = "61-80%",
   "V" = "81-100%"
 )
+usethis::use_data(freqClasses_numToPerc, overwrite = TRUE, internal = TRUE)
 
 freqClasses_numToName <- list(
   "I" = "Scarce",
@@ -76,12 +80,14 @@ freqClasses_numToName <- list(
   "IV" = "Constant",
   "V" = "Contant"
 )
+usethis::use_data(freqClasses_numToName, overwrite = TRUE, internal = TRUE)
 
 # Cover Method Options ----------------------------------------------------
 coverMethod_options <- list(
   "Direct Percentage" = "directPercentage",
   "Domin Class" = "dominCover"
 )
+usethis::use_data(coverMethod_options, overwrite = TRUE, internal = TRUE)
 
 # Habitat Restriction Options ---------------------------------------------
 habitatRestriction_options <- list(
@@ -98,22 +104,27 @@ habitatRestriction_options <- list(
   "Maritime cliff communities (MC)" = "MC",
   "Vegetation of open habitats (OV)" = "OV"
 )
+usethis::use_data(habitatRestriction_options, overwrite = TRUE, internal = TRUE)
 
+# Floristic Tables Options ------------------------------------------------
+floristicTablesView_options <- c("Single Composed vs NVC" = "singleComposedVsNVC",
+                                 "Multiple Composed" = "multipleComposed")
+usethis::use_data(floristicTablesView_options, overwrite = TRUE, internal = TRUE)
 
-# Cross-tabulate NVC floristic tables options -----------------------------
+# Initialise the set of floristic tables to view, will be updated in sidebar_server
+floristicTablesSetView_options <- c("all")
+usethis::use_data(floristicTablesSetView_options, overwrite = TRUE, internal = TRUE)
+
 matchSpecies_options <- c("No" = "No",
                           "Composed to NVC" = "compToNVC",
                           "NVC to Composed" = "NVCToComp")
-
+usethis::use_data(matchSpecies_options, overwrite = TRUE, internal = TRUE)
 
 # Results to View NVC Assignment ------------------------------------------
-resultsViewNVCAssign_options <- c(#"Site, Jaccard" = "nvcAssignSiteJaccard",
-                                  #"Group, Jaccard" = "nvcAssignGroupJaccard",
-                                  "Site, Czekanowski" = "nvcAssignSiteCzekanowski",
+resultsViewNVCAssign_options <- c("Site, Czekanowski" = "nvcAssignSiteCzekanowski",
                                   "Group, Czekanowski" = "nvcAssignGroupCzekanowski",
-                                  "Quadrat, Jaccard" = "nvcAssignPlotJaccard"#,
-                                  # "Quadrat, Czekanowski" = "nvcAssignQuadratCzekanowski"
-                                  )
+                                  "Quadrat, Jaccard" = "nvcAssignPlotJaccard")
+usethis::use_data(resultsViewNVCAssign_options, overwrite = TRUE, internal = TRUE)
 
 # Results to View EIVs ----------------------------------------------------
 resultsViewEIVs_options <- c("Weighted Mean Hill-Ellenberg Values, by Site" = "weightedMeanHEValuesSite",
@@ -121,9 +132,8 @@ resultsViewEIVs_options <- c("Weighted Mean Hill-Ellenberg Values, by Site" = "w
                              "Weighted Mean Hill-Ellenberg Values, by Group" = "weightedMeanHEValuesGroup",
                              "Unweighted Mean Hill-Ellenberg Values, by Group" = "unweightedMeanHEValuesGroup",
                              "Weighted Mean Hill-Ellenberg Values, by Quadrat" = "weightedMeanHEValuesQuadrat",
-                             "Unweighted Mean Hill-Ellenberg Values, by Quadrat" = "unweightedMeanHEValuesQuadrat"
-                             )
-
+                             "Unweighted Mean Hill-Ellenberg Values, by Quadrat" = "unweightedMeanHEValuesQuadrat")
+usethis::use_data(resultsViewEIVs_options, overwrite = TRUE, internal = TRUE)
 
 # Results to View Diversity -----------------------------------------------
 resultsViewDiversity_options <- c("Site Summary Table"  = "diversitySummaryTable",
@@ -131,14 +141,13 @@ resultsViewDiversity_options <- c("Site Summary Table"  = "diversitySummaryTable
                                   "Species Richness, by Site" = "speciesRichnessSite",
                                   "Species Richness, by Group" = "speciesRichnessGroup",
                                   "Species Richness, by Quadrat" = "speciesRichnessQuadrat")
-
-
+usethis::use_data(resultsViewDiversity_options, overwrite = TRUE, internal = TRUE)
 
 # DCA Axis Selection Options ----------------------------------------------
 dcaAxisSelection_options <- c("DCA1 vs DCA2" = "dca1dca2",
                               "DCA1 vs DCA3" = "dca1dca3",
                               "DCA2 vs DCA3" = "dca2dca3")
-
+usethis::use_data(dcaAxisSelection_options, overwrite = TRUE, internal = TRUE)
 
 # DCA Variable Options To Show --------------------------------------------
 dcaVars_options <- c("Survey Quadrats" = "surveyQuadrats",
@@ -149,23 +158,20 @@ dcaVars_options <- c("Survey Quadrats" = "surveyQuadrats",
                      "Unique Survey Species" = "uniqSurveySpecies",
                      "Hill-Ellenberg" = "hillEllenberg",
                      "Survey Quadrat Change" = "surveyQuadratChange")
-
+usethis::use_data(dcaVars_options, overwrite = TRUE, internal = TRUE)
 
 # CCA options -------------------------------------------------------------
 ccaVars_vals <- list("FN" = c("F", "N"),
                      "FR" = c("F", "R"),
                      "FL" = c("F", "L"),
                      "FS" = c("F", "S"),
-                     
                      "NR" = c("N", "R"),
                      "NL" = c("N", "L"),
                      "NS" = c("N", "S"),
-                     
                      "RL" = c("R", "L"),
                      "RS" = c("R", "S"),
-                     
-                     "LS" = c("L", "S")
-                     )
+                     "LS" = c("L", "S"))
+usethis::use_data(ccaVars_vals, overwrite = TRUE, internal = TRUE)
 
 ccaVars_options <- c("Moisture (F) x Nitrogen (N)" = "FN",
                      "Moisture (F) x Reaction (R)" = "FR",
@@ -176,38 +182,37 @@ ccaVars_options <- c("Moisture (F) x Nitrogen (N)" = "FN",
                      "Nitrogen (N) x Salinity (S)" = "NS",
                      "Reaction (R) x Light (L)" = "RL",
                      "Reaction (R) x Salinity (S)" = "RS",
-                     "Light (L) x Salinity (S)" = "LS"
-                     )
-
-
+                     "Light (L) x Salinity (S)" = "LS")
+usethis::use_data(ccaVars_options, overwrite = TRUE, internal = TRUE)
 
 # Global Reference DCA Space Options --------------------------------------
-nationalReferenceSpaces_options <- sort(c(c("A", "CG", "H", "M", "MC", "MG", "OV", "S", "SD", "SM", "U", "W"), setdiff(readRDS(file = "./data/nvc_community_codes.rds"), c("SM1", "SM1a", "SM1b"))))
-
+nationalReferenceSpaces_options <- sort(c(c("A", "CG", "H", "M", "MC", "MG", "OV", "S", "SD", "SM", "U", "W"), setdiff(readRDS(file = "./inst/extdata/nvc_community_codes.rds"), c("SM1", "SM1a", "SM1b"))))
+usethis::use_data(nationalReferenceSpaces_options, overwrite = TRUE, internal = TRUE)
 
 # DCA Survey Quadrat Options ----------------------------------------------
 surveyQuadratSelection_options <- c("All" = "all",
                                     "Select Years" = "selectYears",
                                     "Select Groups" = "selectGroups",
                                     "Select Quadrats" = "selectQuadrats")
+usethis::use_data(surveyQuadratSelection_options, overwrite = TRUE, internal = TRUE)
 
 selectSurveyYears_options <- c()
+usethis::use_data(selectSurveyYears_options, overwrite = TRUE, internal = TRUE)
 
 selectSurveyQuadrats_options <- c()
+usethis::use_data(selectSurveyQuadrats_options, overwrite = TRUE, internal = TRUE)
 
 selectSurveyGroups_options <- c()
-
+usethis::use_data(selectSurveyGroups_options, overwrite = TRUE, internal = TRUE)
 
 # Group Survey Plots Options ----------------------------------------------
 groupSurveyPlots_options <- c("No" = "no",
                               "Group" = "group",
                               "Year" = "year")
+usethis::use_data(groupSurveyPlots_options, overwrite = TRUE, internal = TRUE)
 
 # Report Options ----------------------------------------------------------
-reportOptions_options <- list(#`NVC Assignment Pseudo-quadrat` = c("Site" = "nvcAssignmentResultsSite",
-                              #                                    "Group" = "nvcAssignmentResultsGroup",
-                              #                                    "Quadrat, Jaccard" = "nvcAssignmentResultsQuadrat_Jaccard"),
-                              `NVC Assignment` = c("Site, Czekanowski" = "nvcAssignmentResultsSite_Czekanowski",
+reportOptions_options <- list(`NVC Assignment` = c("Site, Czekanowski" = "nvcAssignmentResultsSite_Czekanowski",
                                                    "Group, Czekanowski" = "nvcAssignmentResultsGroup_Czekanowski",
                                                    "Quadrat, Jaccard" = "nvcAssignmentResultsQuadrat_Jaccard"),
                               `Floristic Tables` = c("Site" = "composedFloristicTablesSite",
@@ -229,4 +234,4 @@ reportOptions_options <- list(#`NVC Assignment Pseudo-quadrat` = c("Site" = "nvc
                                         "Local (unrestricted)" = "mvaLocalReferenceUnrestricted"),
                               `Survey Table` = c("Survey Table" = "surveyData")
     )
-
+usethis::use_data(reportOptions_options, overwrite = TRUE, internal = TRUE)