+++ modules/calcAvgEIVs_server.R
@@ -288,7 +288,7 @@ calcAvgEIVs <- function(input, output, session, surveyData, sidebar_options) {
       # By Quadrat
       weightedMeanHEValuesQuadrat <- surveyData_long |>
         dplyr::rename("species" = "Species") |>
-        dplyr::left_join(master_data, by = "species",
+        dplyr::left_join(RMAVIS::master_data, by = "species",
                          relationship = "many-to-many") |>
         dplyr::select(Year, Group, Quadrat, Cover, `F`, L, N, R, S) |>
         dplyr::mutate("F" = `F` * Cover,
@@ -343,7 +343,7 @@ calcAvgEIVs <- function(input, output, session, surveyData, sidebar_options) {
       # By Quadrat
       unweightedMeanHEValuesQuadrat <- surveyData_long |>
         dplyr::rename("species" = "Species") |>
-        dplyr::left_join(master_data, by = "species",
+        dplyr::left_join(RMAVIS::master_data, by = "species",
                          relationship = "many-to-many") |>
         dplyr::select(Year, Group, Quadrat, Cover, `F`, L, N, R, S) |>
         dplyr::group_by(Year, Group, Quadrat) |>