+++ modules/floristicTables_ui.R
@@ -3,28 +3,58 @@ floristicTablesUI <- function(id){
   ns <- NS(id)
   
   shiny::fluidRow(
-    shiny::column(
-      width = 6,
+    
+    shiny::div(
       
-      # shiny::h5("Composed Floristic Table"),
-      shiny::htmlOutput(outputId = ns("composedFloristicTableTitle")),
+      id = ns("singleComposedVsNVC_div"),
       
-      shiny::div(
-        reactable::reactableOutput(outputId = ns("floristicTables_composed"))
+      shiny::fluidRow(
+        
+        shiny::column(
+          width = 6,
+          
+          # shiny::h5("Composed Floristic Table"),
+          shiny::htmlOutput(outputId = ns("composedFloristicTableTitle")),
+          
+          shiny::div(
+            reactable::reactableOutput(outputId = ns("floristicTables_composed"))
+          )
+          
+        ),
+        shiny::column(
+          width = 6,
+          
+          # shiny::h5("NVC Floristic Table"),
+          shiny::htmlOutput(outputId = ns("nvcFloristicTableTitle")),
+          
+          shiny::div(
+            reactable::reactableOutput(outputId = ns("floristicTables_nvc"))
+          )
+          
+        )
+        
       )
       
     ),
-    shiny::column(
-      width = 6,
+    
+    shiny::div(
       
-      # shiny::h5("NVC Floristic Table"),
-      shiny::htmlOutput(outputId = ns("nvcFloristicTableTitle")),
+      id = ns("multipleComposed_div"),
       
-      shiny::div(
-        reactable::reactableOutput(outputId = ns("floristicTables_nvc"))
+      shiny::column(
+        width = 12,
+        
+        shiny::htmlOutput(outputId = ns("floristicTablesWide_composedTitle")),
+        
+        shiny::div(
+          reactable::reactableOutput(outputId = ns("floristicTablesWide_composed"))
+        )
+        
       )
       
+      
     )
+    
   )
-  
+    
 }