+++ tests/testthat/test-read_mavis_data.R
@@ -1,0 +1,6 @@
+testthat::test_that("read_mavis_data works", {
+  
+  
+  
+  testthat::expect_equal(2 * 2, 4)
+})