+++ NEWS.md
@@ -1,3 +1,107 @@
-# RMAVIS (development version)
+---
+title: "Release notes for `RMAVIS`"
+output: html_document
+---
 
-* Initial CRAN submission.
+# RMAVIS v0.983 (2024/03/14)
+*   Modifications:
+      -   Converting to R package format.
+*   Fixes:
+      -   Fixing MVA select quadrats.
+*   New Features:
+      -   Adding multiple composed floristic tables.
+
+# RMAVIS v0.982 (2024/03/11)
+
+*   Modifications: NA
+*   Fixes:
+      -   Fixing Species.Ignore == "Yes"
+*   New Features: NA
+
+# RMAVIS v0.981 (2024/03/09)
+
+*   Modifications: NA
+*   Fixes:
+      -   Ensuring species-quadrat duplicate check is included in ok to proceed.
+      -   Fixing jump issue in species adjustment table, now using Yes and No.
+*   New Features: NA
+
+# RMAVIS v0.98 (2024/03/06)
+
+*   Modifications: NA
+*   Fixes:
+      -   Ensuring manual entry works.
+      -   Ensuring single plot samples work.
+*   New Features: NA
+
+# RMAVIS v0.97 (2024/03/05)
+
+*   Modifications: NA
+*   Fixes:
+      -   Updating documentation to reflect new MVA functionality.
+      -   Fixing group survey plots option in MVA module.
+*   New Features: NA
+
+# RMAVIS v0.96 (2024/03/02)
+
+*   Modifications:
+    -   Changing name, MAVIS -> RMAVIS
+    -   Removing site and group Jaccard similarity results.
+    -   Refining code throughout.
+    -   Trimming dependencies.
+*   Fixes: NA
+*   New Features:
+    -   Adding pseudo-quadrat and sample quadrat centroids in MVA.
+    -   Adding setupData module which allows modification of bundled datasets.
+    -   Adding a bryophyte switch to include/exclude bryophyte taxa.
+    -   Now loading dependencies in `load_packages.R`.
+
+# RMAVIS v0.95 (2024/01/24)
+
+*   Modifications:
+    -   Removing unnecessary upload data checks.
+    -   Making survey data availability table searchable.
+    -   Changing nomenclature for releve to plot.
+*   Fixes:
+    -   Ensuring report renders correctly using new surveyDataSummary
+        object data.
+*   New Features: NA
+
+# RMAVIS v0.94 (2024/01/18)
+
+*   Modifications: NA
+*   Fixes:
+    -   Ensuring the number of quadrats per group is passed to the
+        floristic tables module through the new survey data summary
+        module.
+*   New Features: NA
+
+# RMAVIS v0.93 (2024/01/18)
+
+*   Modifications: NA
+*   Fixes: NA
+*   New Features:
+    -   Adding survey data summary module.
+    -   Adding quadrat re-grouping tool.
+
+# RMAVIS v0.92 (2024/01/18)
+
+*   Modifications: NA
+*   Fixes: NA
+*   New Features:
+    -   Adding Czekanowski index method for Site and Group NVC
+        assignment calculations.
+
+# RMAVIS v0.91 (2024/01/11)
+
+*   Modifications: NA
+*   Fixes: NA
+*   New Features:
+    -   Adding privacy policy section
+
+# RMAVIS v0.9 (2024/01/09)
+
+*   First release for user testing.
+*   Modifications: NA
+*   Fixes: NA
+*   New Features: NA